Added barebones lobby logic to start menu
