--- weave/Menus/LobbyDemo.tscn
@@ -1,30 +1,0 @@
-[gd_scene load_steps=2 format=3 uid="uid://suha7hqfnwef"]
-
-[ext_resource type="Script" path="res://Scripts/MenuControllers/LobbyDemo.cs" id="1_5iwbe"]
-
-[node name="LobbyDemo" type="Control"]
-layout_mode = 3
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-script = ExtResource("1_5iwbe")
-
-[node name="Button" type="Button" parent="."]
-layout_mode = 0
-offset_left = 512.0
-offset_top = 91.0
-offset_right = 553.0
-offset_bottom = 122.0
-text = "play"
-
-[node name="TextEdit" type="TextEdit" parent="."]
-layout_mode = 0
-offset_left = 221.0
-offset_top = 168.0
-offset_right = 908.0
-offset_bottom = 541.0
-mouse_filter = 2
-placeholder_text = "Players here..."
-editable = false
